Change the behavior bin/dev_scripts/update-localization-files script.
With this change the script will only update the webwork2.pot file using
xgettext.pl by default.  It will not also update the .po files using
msgmerge.

However, there are now options for this script.  If you run the script
with no options it will do what is stated above.  If you add the option
`-p`, then it will also update all .po files.  If you instead use the
option `-l` with a language code, then it will update only that .po file
in addition to updating the webwork2.pot file.  For example if you run
`update-localization-files -l fr-CA`, then it will update the
webwork2.pot file, and then update the fr-CA.po file.

The point of this is that with the switch to using Transifex, we no
longer need to directly update po files.  We only need to update the
webwork2.pot file.  When those changes are merged into the develop
branch, then Transifex will automatically get those changes, and those
changes will be added to all of the languages there.  When Transifex
makes a pull request back to webwork2 for a language, those strings will
be included.

The optional flags are for translators.  If a translator wants to add
the current strings from the code to the po file so that they can start
working on translations before we merge the updated webwork2.pot file,
then they can use this script with the options to do so.

This also adds the conf directory to those xgettext.pl searches for
translation strings as is done in #1631 to make sure that gets into this
file in its new location.