Remove the MIME-Header/Base64 encoding of the full name in the email to and from addresses.
It is incorrect to do this for all names regardless of if the name contains characters which neccessitate such encoding and raises the spam score of the email address in cases where the name does not contain such characters.

Email::Stuffer automatically encodes names that do have such characters, and does not do so if the name lacks such characters, so this encoding by webwork is no longer neccessary.