Load OPL set definitions from files instead of searching the OPL directory.
This is done in the getDefList method of lib/WeBWorK/ContentGenerator/Instructor.pm.  The only time that method is called is when the homework sets editor is loaded.  Note that the course's templates directory is still searched, but the Library and Contrib subdirectories are pruned and so not look into.

A new script `bin/generate-OPL-set-def-lists.pl` is created that generates the needed data files by searching the OPL directory.  Note that this script is executed by `bin/OPL-update`.  For now, this script is executed in docker-entrypoint.sh when the docker container is started.  However, these files should be added to the OPL release built by webwork2/bin/OPL_releases/release.sh, and dealt with in the same way as the other files in webwork2/htdocs/DATA.  Of course, that whole process needs to be updated and removed from the personal fork it is currently in.  The release there is now more than two years old.

The defaults.config option $option{useOPLdefFiles} still is checked to determine if OPL set defintion files will be loaded and listed, but note that the $options{setDefSearchDepth} is ignored for this.  The new defaults.config option $option{useContribDefFiles} determines if set definition files from the Contrib directory will be loaded and listed.  This also ignores depth.  So the search depth option only applies to local directories for the course.