Make the select on the file manager page vertically fill the div that it is contained in.
Make the heading of the attepts table an `h2` instead of an `h3`.

Make the `.text-primary` color the same as the link color for the
math4-yellow theme.  The yellow primary color is not accessible for a
text color.  This can be seen on the library browser page once problems
are loaded.  The "Global Usage", "Attempts", and "Status" are all in
this color.

On the users assigned to set page, change the message from bad to good
when the set is assigned to all users.  That is what is used to be and
should be.