Delete extra tables and fields when a course database is upgraded.
If a course has tables or fields that are not present in the database schema, then a checkbox will be displayed below the table or field asking to delete the table or field when the course is upgraded.

In order to make it easy to test this, the obsolete `facebooker` column is deleted from the global_user_achievement table (and all related code removed).  That does not work.  I personally don't think that integrating with a social media platform is a good idea for WeBWorK in any case.

Note that the code previously added to delete the old "published" field has been removed.  If you have that column, then you can now upgrade the course, and check to delete that column.  This is a more correct way to deal with this.  I believe this fixes issue #1419.