Breadcrumb link tweaks (issue #1660)
In the breadcrumb make the problem number not a link on the problem page.  That is a link to the page the user is currently on, and so should not be a link.  This is also consistent with the breadcrumb for other pages.

Also add "Show Me Another" to the breadcrumb on the ShowMeAnother page.