Graded proctored gateway quizzes are no longer proctored
On the ProblemSet.pm page for a proctored gateway quiz if there is a
version that has been submitted and graded, it should no longer use the
proctored gateway quiz module.  It should use the usual gateway quiz
module.  The mistake was in the assumption that all versions are of the
same type as the template.  This is an incorrect assumption.