Fix the hasPermission calls for the navigation_allowed permission in Authen.pm.
When the proctored gateway quiz module is used and proctor authentication is obtained, the userID string that is passed into create_session is the student user id and the proctor user id separated by a comma.  In addition, when the gateway quiz is graded and proctor authentication is obtained in that case ",g" is appended to the end ofthat.  So those will not work in the hasPermission call.  This strips off the proctor user id and ",g" to pass as the user id to the hasPermission call.