In the gateway quiz javascript, choose the last button name submitAnswers.
If there are more than one submitAnswers button on a gateway quiz page,
choose the last one to use for the confirmation dialog and other things
done in the gateway.js file.  This fixes a javascript error caused by
things like the problemPanic.pl macro adding another submitAnswers
button.  It is highly unlikely that the problemPanic.pl macro will work
as expected in a gateway quiz.  This doesn't fix that.

See issue #1699.