Fix when a proctored gateway quiz is changed to a regular gateway quiz.
There were two manifestations of the issue here.

First, if a professor views an in progress proctored gateway quiz when acting as a student and does not have the permission to record_answers_when_acting_as_student then the quiz should obviously remain a proctored gateway quiz.  Currently a simply the act of viewing a proctored quiz when acting as a student will change a proctored gateway quiz into an unproctored one.  This is clearly really bad.  This means that the student may now continue to work the quiz without needing proctor authentication to grade the quiz.

Second, if a professor does have the permission to record_answers_when_acting_as_student, grades the quiz, and uses the last attempt at a quiz in doing so, then the proctored quiz should be changed into an unproctored quiz.  Currently, that is not done.  This means that the student will not be able to continue to work the quiz, but also will need to obtain proctor authentication to review the quiz.

Note that both of these bugs were present in previous versions of webwork.

Also, note that this includes both #1698 and #1697.  This is because without both of those you can not properly test this pull request as they both interfere with proper behavior of a proctored gateway quiz.