Remove HTML from page title
The URLPath name method can not contain a span with the dir="ltr"
attribute when it is used in the page title. However, it needs that
markup when it is displayed as a page heading for left-to-right
languages like Hebrew.  So a new $displayHTML argument is added to the
method that tells it if that is desired.

This argument is then passed for the ContentGenerator.pm title method
which is used for the system.template page-title that displayed as a
page heading.

The ContentGenerator.pm `title` method is used for what is called
"page-title" in the system.template, but is not the "page title".  The
ContentGenerator.pm `path` method is use for the page title.  Both call
the URLPath name method.

This fixes issue #1703.