Fix the handling of moving/removing the temporary course archive file.
In the process of creating a course archive, first a temporary tar file
is created in the courses directory.  Then that file is moved to the
requested location.

There were several issues with the coding of how the temporary file was
moved to the requested location.

First the perl `rename` method was used.  If the requested location is
not on the same file system as the courses directory that will fail.  So
this switches to using the `move` method in the `File::Copy` module.
Note that webwork already uses this module elsewher, so this is not a
new dependency.

Second, in the case that the `rename` method fails, the error was never
checked.  The result was that the temporary archive file is left in the
courses directory, and no errors are reported even though the requested
archive file does not exist.

Third, in the case that a previous archive file exists in the courses
directory, the `croak` call occured before the `unlink` call to delete
the temporary archive file (in the clean up according to the comments).
Thus the temporary archive file is not actually deleted.

This also removes the space at the beginning of the temporary file name.
That worked, but why have it?