Add the package-lock.json file back.
It seems that bootstrap version 5.2.0 was recently released and is not
entirely compatible with version 5.1.3.  This causes the bootstrap.scss
file to fail to compile.

This adds a package-lock.json file from before the version update
occured to lock us in to 5.1.3.

Another way to fix this is to import the
`node_modules/bootstrap/scss/maps` file in our bootstrap.scss file.
However, that may only be a temporary fix as bootstrap may update again
and change something else that breaks our build.

Now that we are also insisting on everyone using node 16 this should not
cause the problem that we had before.