Move the WeBWorK::PG module to PG.
Note that the WeBWorK::PG::Local and WeBWorK::PG::Remote modules have
been removed.  The Local variant is integrated directly into
WeBWorK::PG, and the Remote variant was not working and is in disrepair.

Remove the pg special environment variables
ALWAYS_SHOW_HINT_PERMISSION_LEVEL and
ALWAYS_SHOW_SOLUTION_PERMISSION_LEVEL.  Use the always_show_hint and
always_show_solution permissions directly.  Whether solutions and hints
are shown is now controlled by webwork2 and not by pg.

The pg variable/flag $showHint is removed.  When and if hints are shown
is now controlled by the instructor with a combination of options at the
course, set, and problem level.

Note that hints were not implemented in gateway quizzes and still
aren't.

Also the options to use knowls for solutions and hints have been
removed.  Always use knowls for solutions and hints.

Showing the various pieces of debugging information for a problem is now
controlled with translation options.  Webwork permissions are converted
into these debugging options.  The VIEW_PROBLEM_DEBUGGING_INFO special
pg environment variable was removed and is one of these debugging
options.

Although the permissionLevel and effectivePermissionLevel are still
passed in, they are no longer used by anything in PG.  A translator
option named isInstructor is passed into the environment that can be set
to get instructor things to happen.  Another special translator option
for scaffolds named forceScaffoldsOpen can be passed to force all
scaffolds to be open.  This is set in the library browser and in gateway
quizzes.

Reduced scoring is implemented on the webwork2 side.  PG no longer does
that.  It is implemented with the old method where the reduced score is
saved in the problem status and the score before the reduced scoring
period begins is saved in the sub_status.  This can be changed later to
fix some of the other issues with reduced scoring.

Several other environment variables that are clearly not needed have
also been removed including the appletPath and session key.

Note that PGalias no longer uses the course name, user login, set
number, and problem number for generating resource UUID's.  It now only
uses the problem seed, psvn, and problemUUID.  So webwork2 constructs a
problemUUID that guarantees the necessary uniqueness.