On the grades page only hide scores for gateway sets.
The code that checks to see if scores for a gateway quiz should be
hidden on the grades page was not checking to see if the set was
actually a gateway quiz.  However, it is possible to make a homework set
have hide_score set to 'Y'.  To make this happen create a homework set,
change it to a gateway/quiz, and then set it to not show scores on
finished assignments and save.  Then change the assignment type back to
homework and save (only once).  Note that if you click save one more
time it will reset the hide_score value in the database back to NULL.

If a student is assigned one of these homework sets and tries to view
the grades page it causes an error.

See https://webwork.maa.org/moodle/mod/forum/discuss.php?d=8092.