Drop the CSS override forcing CodeMirror to have the LTR direction. That is a default of the version of CodeMirror now being used in WeBWorK (loaded via the "npm install").
Modify the value of unicode-bidi for CodeMirror lines, as BootStrap sets
       unicode-bidi: bidi-override;
in node_modules/bootstrap/scss/_reboot.scss for all HTML pre elements,
and that gets applied to the elements inside the CodeMirror editor also.
The BootStrap settng interferes with the Unicode bidirectional algorithm
processing Hebrew (and other right-to-left language) text inside the lines
in the CodeMirror editor, so that the characters appear from left to right
(logical layout) instead of from right-to-left (the expected visual
layout).