Quick hack to allow the gateway quiz page to be viewed with restricted IP settings.
This allows the new gateway quiz summary page to be viewed even if IP
restrictions don't allow access to the gateway quiz itself.  The gateway
quiz module prevents the quiz itself to be entered still.  This is a
somewhat ugly hack though.  It checks the return value of checkSet to
see if it matches /^Client ip address .* is not in the list of
addresses/ and allows gateway quiz sets through if it is.

This also fixes a call in Hardcopy.pm to the now non-existent
Apache2::Connection remote_ip method.  That needs to be the
Apache2::Request useragent_ip method now.  Apparently this has been
broken since the Apache 2.4 methods were added.  In Authz.pm this change
was made, but it must have been missed here.