Replace mod_perl2 with Mojolicious.
The WeBWorK::Cookie module which was really just an alias for the
CGI::Cookie module has been removed.  Instead this just uses the
Mojolicious cookie methods.  Note that this does change the way the
$CookieLifeTime setting works.  Previously it was a string
representation of the life time like "+7d", and now it must be numeric
and is always in seconds or the string "session" (in which case the
expires and max-age values are not set and the cookie expires with the
session).

Note that configuration that was previously in
webwork2.apache2.4.dist.conf is now in webwork2.mojolicious.dist.yml.
Unless you want to change the settings in that file, you don't need to
copy that file to webwork2.mojolicious.yml.  The `dist` file will be
used directly if the webwork2.mojolicious.yml file does not exist.