Replace the XMLRPC::Lite and Apache::XMLRPC::Lite usage with a direct remote procedure call endpoint.  This removes the XMLRPC middle man.
Note that the paths for this have changed.

The html2xml path still exists, but is an alias for the new render_rpc
path.  The html2xml name doesn't make sense anymore since it does not
use xml at all.  It should be considered deprecated and not used in new
code.  It is provided to give those that use it time to make the change.
It still works the same at it did before.

The instructorXMLHandler path is now instructor_rpc.  Note that the
response that is returned is now JSON, instead of a string that is
parsed into JSON. That is unless there is an error, in which case it is
text.  There is no backwards compatibility path for this.  I don't think
that this path was really used by anyone.  If I am mistaken, an alias
can be made.

Note that all webwork2 calls have been updated to the new paths.

Note that the badly broken, previously disabled, and probably never
completely written WebworkWebservice::MathTranslators package has been
deleted.  In addition many of the methods in the other WebworkWebservice
packages now actually work.

Also a global javascript `webworkConfig` variable is now injected into
the templates (in system.template and gateway.template) that provides
system information for javascript files.  At this point it contains the
single key `webwork_url`.  This is the value of $webwork_url in
site.conf (typically /webwork2).  Although it is actually the empty
string if $webwork_url is set to '/'.  This makes that setting actually
work again.