Add a script for running the morbo webserver for webwork2.
The script takes care of adding some watch directories.  It really is
the morbo script, so it accepts all of the options that morbo does.  You
don't need to pass the webwork2 app script as that is automatically
found.

Run the script with ./bin/dev_scripts/webwork2-morbo from the webwork2
root directory (although it will work from other locations too).