Fix an issue in gateway quizzes that causes correct answers to be displayed on all submissions.
The removal of the $pg{options}{showCorrectAnswers} from defaults.config
made it so that the $want->{showProblemGrader} value in GatewayQuiz.pm
could possibly be undefined.  As a result the AttemptsTable.pm code
defaulted to showing correct answers.  Both of those are changed.  A
fallback was added so that the $want->{showProblemGrader} is now defined
and 0 in this case.  Although that isn't really needed.  Why did the
AttemptsTable.pm default to displaying correct answers if the
showCorrectAnswers option was undefined?  That doesn't make sense.  So
that default is changed.