Fix more issues with the gateway problem numbers.
There were still cases where non-consecutively ordered gateway quizzes
can cause errors.  In fact any time that a gateway quiz does not have a
problem number 1 in the set definition, you will get an error.  In
addition answer processing had issues that could cause errors.
Basically, arrays storing problem data should generally not be accessed
by the direct numerical index, but indirectly via the @probOrder array
index.  There are some exceptions to this.  One exception is noted.

Note that this also fixes an issue with randomly ordered gateway
quizzes.  The issue is perhaps minor, but the data stored in a column of
the past_answers table was disjoint.  The past_answers_string and scores
did not match up with the problem_id and source_file.  The source_file
in that table is not used anywhere by webwork, but it still should
match.  The problem_id is used.  In the past answers table, this
mismatch leads to some rather unexpected results.  If a student takes
multiple versions of the quiz, and you view the answers for all versions
for that student, the answers clearly look like they come from different
problems.  The same happens if you view answers for multiple students.
One downside of this is that now the answers are not in the order that
the problems appear for the student.  They are in the order they appear
in the set definition.  That is consistent for all students and
quiz versions though.