Disable browser password autocompletion on the proctor login page.
This just adds the attribute `autocomplete="new-password"` to the
proctor user and password fields.  This is the recommended way to do
this according to
https://developer.mozilla.org/en-US/docs/Web/Security/Securing_your_site/Turning_off_form_autocompletion#preventing_autofilling_with_autocompletenew-password.

However, this may not be supported by all browsers.  Safari is listed as
support unknown.  So could a Safari user test this and let me know if it
works?  If it doesn't, I have another way that should work for all
browsers.