Fix the showHintsAfter setting for old style problem list files
I didn't test this for version 1 type problem list set definition files
when I added the showHintsAfter setting, and of course it doesn't work.
The showHintsAfter setting is guaranteed to not be in any of these
files, and so the setting ends up being set to the empty string.  The
old style list format doesn't fill in defaults, and so the empty string
is put into the database.  The database interprets this as a null value
and croaks since the column is not null.

To fix this the method for parsing the old style lists just needs to use
the default value from the conf files.