Improvements for course info files.
The save_as_handler is updated to handle course info files correctly.
I don't think this ever worked right.

If a course information file does not exist the "New Version" tab is
shown but the file name input is readonly.  The course info file name
needs to be what is set in the course environment for that.  In addition
the "Update" tab is not shown.

If a course information file does exist the "New Version" tab is not
shown but the "Update" tab is.

Also, if the course info file does not exist then don't show the
'Revert' tab in any case.  There is nothing to revert to.

The javascript is updated to handle things for course info files better
also.