Modify hypnotoad configuration and add more documentation.
Set clients to 1 as WeBWorK is blocking code, increase the
default value of graceful_timeout (to further reduce the risk
of trouble from hard kills of workers) and raise the default
value of workers to allow handling more requests in parallel.

Modify several other settings.

Remark about the need for production systems to tune the value
of workers and spare.