Dispatch with the Mojolicious router.
Make all ContentGenerator modules actual controllers and use the
Mojolicous router for dispatching requests.  The WeBWorK::Request module
is now the WeBWorK::Controller module. It derives from the
Mojolicious::Controller package, and is that base class for all WeBWorK
Mojolicious controller modules.  The WeBWorK::ContentGenerator module
derives from that, and as before all other content generator modules
derive from that.  So the package hierarchy is:
	Mojolicious::Controller
	-> WeBWorK::Controller
	-> WeBWorK::ContentGenerator
	-> WeBWorK::ContentGenerator::...
In the ContentGenerator modules the old $r variable (which stood for an
Apache::Request) has been replaced with $c (for
Mojolicious::Controller).  Furthermore, $self is gone.  There is only
$c.  The old $self and $r are now both $c.  The current controller
module is available in the templates (made available by Mojolicious) as
$c, so this naming is consistent bot in the module itself and in the
templates.  Yeah, $self is also made available in the templates by
Mojolicious and $c == $self, but $c is shorter and this is consistent
with the Mojolicious documentation also.

Note that signatures have been enabled in all content generator modules
(and a few other modules).

lib/WeBWorK.pm still initializes all ContentGenerator requests by
initializing the course environment and checking authentication, but
then it relinquishes control to the appropriate ContentGenerator
controller module.  That is except for the case when the user is not
authenticate, in which case the Login.pm module is rendered directly,
and the case that login proctor authentication is needed, in which case
the LoginProctor.pm module is rendered directly.  This is done in the
`around_action` hook.

Mojolicious routes are initialized in the WeBWork::Utils::Routes module.
That module also has two convenience methods that return webwork
specific parameters that are not Mojo::Routes:Route properties.  Those
are the route_title which returns the translated display title for a
route, and route_navigation_is_restricted, which determines if a route
is allowed for LTI navigation restricted users.  Note that the
WeBWorK::URLPath module is no more.  Everything that did can be done
directly with the Mojo::Routes or Mojo::Routes::Route modules or the two
convenience methods of the WeBWork::Utils::Routes module.

The most prominently utilized Mojolicious method that replaces the
old WeBWorK::URLPath usage is the `url_for` Mojolicious::Controller
method.  This returns a Mojo::URL.  It automatically incorporates stash
values (previously urlpath args) for the current route into the
generated url.  Furthermore, the Mojo::URL `query` method is used to
simplify adding query parameters.  The WeBWorK::ContentGenerator
`systemLink` method is rewritten to utilize this.  That modules
`url_args` and `url_authen_args` methods are revised to not directly
generate the tail of the url, but to return a hash that can be passed to
the `query` method.

Note that the WeBWorK::FakeRequest module has been deleted.  That is not
only not used, but also won't work anymore.  It could be made to work
again, but it would take some effort and we really don't need it anymore.