Set definition file handling improvements.
The set definition utility methods in ProblemSetList.pm have been moved
into a separate file.  The new file is `lib/WeBWorK/File/SetDef.pm`.
The methods in this file do the same things that the previous methods in
`lib/WeBWorK/ContentGenerator/Instructor/ProblemSetList.pm` with some
improvements. This separates the set definition file handling from the
content generator user interface code.

The `readSetDef` no long returns a reference to a messy array of
disorganized data.  Instead it returns a reference to a hash whose keys
give information as to what the data fields are. Using this hash in the
method itself also facilitates a much cleaner way to extract the data
from a set definition file than the massive `if .. elsif ..` mess from
before.

Note that if dates are out of order, instead of just leaving them and
warning about it, the defaults relative to the due date are used just as
they would be when creating a new set.  If the due date is not defined,
then one week from the current time be used.  The messages about these
dates being out of order in the file are still shown though.

Another big improvement is that `warn` is never used in the file.
Instead any of those previous warnings are returned in a reference to an
array.  Each element of the array is also an array reference whose
contents are suitable for passing to maketext.  This is one of the
needed steps toward removing the global warning handler.

To accommodate more complex warning messages the `addgoodmessage` and
`addbadmessage` methods need to be placed in a `div` container with the
"alert" role instead of a `p`.  Visibly this makes no difference.  Using
this the warnings from before are now in the alert messages returned
from the SetDef.pm methods.