Update third party libraries to their latest versions.
Well, that is except for codemirror.  Upgrading to version 6 will
require a build system.

The upgrade to Bootstrap 5.2.3 requires a few changes to the code.

The changes to regain the dark borders on table thead sections are
perhaps not necessary.  Those borders were not there before the upgrade
to the previous version of Bootstrap 5.  The previous version forced
those borders.  It is a new feature that those borders are opt in.  This
has to be done per table.  However, I have become accustomed to those
borders, so I opted in.  If you feel differently, let me know.

In order for the accordion buttons on the ProblemSetDetail.pm page to
show up the containing div needs to have the accordion class now.
Fortunately with that everything works as it should.