Implement LTI 1.3 authentication and LTI Advantage tools.
Everything is set up to work much like the old LTI 1.1 authentication
and the previous grade passback.  There is much more that can be done
with the LTI Advantage tools and that could yet be implemented.

There is some hackery so that the courseID can be reliably retrieved
early in the LTI advantage launch request before the JWT is decrypted.

There is a new column added to the set_user (UserSet) table. Other than
that the existing lis_source_did columns from other tables were recycled
for this.  The new column is name lis_source_did the same as the
set_user column so that the merged set method still works without adding
a new column to that table.  It is the LMS lineitem that is stored
there, and that is common to all users of the set.  Note that there are
several new things stored in the course's "setting" table for this, but
that doesn't take a change to the database.

The authen_LTI configuration files are reworked to relegate the two
versions being used at the same time.