Fix the Library Browser add all button.
The problem is that I forgot to add the `type="button"` attribute to the
"Add All" input.  This does better though.  Rather than use an input for
something that is handled by javascript use and actual button.  The
reason an input (with type button) was used before is because CGI didn't
have a real html button.

This fixes issue #1870.