Remove the course_id from the past_answer table.
It is time for this to go. It is an entirely unnecessary column. This is
taken care of by the usual course upgrade procedure in the admin course.

Don't test this on a course that you want to go back to other branches
with! It isn't so easy to add the course_id column back. It can be
done though if needed.