Add commented out lines to enable show-source.cgi to the apache2 config
When proxying the webwork2 app via apache2 these lines can be
uncommented and a link made to htdocs/show-source.cgi in the courses
html directory to enable the "Show Problem Source" button that is added
to a problem by loading the source.pl macro.

The show-source.cgi script is updated to also search for macros in the
new pg/macros subdirectories. It is also perltidy-ed, strictures and
warnings enabled, and cleaned up.