Use Mojolicious::Plugin::SetUserGroup instead of authbind
Instead of using authbind use Mojolicious::Plugin::SetUserGroup.  This
will work on POSIX compliant systems.  According to the documentation
for the plugin that includes Linux, OSX, and BSD (although authbind does
to).  Note that you may need to force install the module as a test is
failing (on my system at least).  The module still works without problem
though.

This will only take effect in production mode (for instance when run via
hypnotoad) and when run by the root user.  It works by switching to the
chosen user and group on the next tick after the Mojo::IOLoop starts.
So under hypnotoad all of the workers are not run as root, but as the
chosen user and group.

Note that Mojolicious::Plugin::SetUserGroup is not a required module,
and should not be added to check_modules.  It is only needed if serving
directly, and not when proxying via another web server like apache2 or
nginx.  The app will run fine without the module installed, and will not
give any warnings or errors unless you try to run hypnotoad as root.
You shouldn't do that unless you are serving directly and have the
module installed anyway.